backend/requirements.txt
# FastAPI Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# LangChain (compatible versions)
langchain>=0.1.0
langchain-openai>=0.0.6
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.30
langgraph>=0.0.20

# Embedding & Vector
pymilvus>=2.3.0
chromadb>=0.4.20

# Document Processing
pypdf>=4.0.0
python-docx>=1.0.0
docx2txt>=0.8

# Retrieval
rank-bm25>=0.2.2
jieba>=0.42.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Storage
minio>=7.1.0

# JWT Authentication
PyJWT>=2.8.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
